Put the functions in utils file and remove from the main function
